+++ src/main/java/com/madcamp/yay/domain/user/repository/FriendRepository.java
@@ -11,5 +11,5 @@ public interface FriendRepository extends JpaRepository<Friend, Long> {
     Optional<Friend> findByEmail(String email);
 
     @Query("SELECT f FROM Friend f WHERE f.user.id = :userId AND f.email = :friendEmail" )
-    Optional<Friend> findByUserIdAndFriendEmail(@Param("userId") Long userId, @Param("friendEmail") String friendEmail);
+    Optional<Friend> findByUserIdAndFriendEmail(@Param("userId") Integer userId, @Param("friendEmail") String friendEmail);
 }
